requirement1.txt
# GUI Framework (Tested versions)
PyQt6==6.7.0
PyQt6-Charts==6.7.0
PyQt6-tools==6.7.0

# Malware Detection (Local - No APIs)
pefile==2023.2.7
yara-python==4.5.0
python-magic-bin==0.4.14
capstone==5.0.1

# System Monitoring
psutil==6.0.0
watchdog==4.0.1
WMI==1.5.1

# Encryption & Security
cryptography==43.0.1
argon2-cffi==24.1.0
bcrypt==4.2.0

# Network & Protocols
scapy==2.5.0
netifaces==0.11.0
dnspython==2.6.1

# Data Processing
pandas==2.2.3
numpy==2.1.1
sqlalchemy==2.0.35

# Report Generation
reportlab==4.2.5
jinja2==3.1.4

# Hardware Control
pyusb==1.2.1
bleak==0.22.3  # Bluetooth LE (Modern replacement)

# Utilities
python-dotenv==1.0.1
colorama==0.4.6

# Core Scanner Engine
yara-python==4.5.1
pefile==2023.2.7
python-magic==0.4.27
psutil==6.0.0

# GUI Framework
PyQt6==6.6.0
PyQt6-Charts==6.6.0
PyQt6-WebEngine==6.6.0

# Security Features
cryptography==43.0.1
argon2-cffi==23.1.0
qrcode[pil]==7.4.2

# Database & Analytics
sqlite3  # Python built-in
pandas==2.2.3
matplotlib==3.9.2

# Network & VPN (Future)
requests==2.32.3
scapy==2.5.0

# Utilities
pathlib  # Python built-in
watchdog==5.0.2
Pillow==11.1.0

# Development & Testing
pytest==8.3.3
black==24.8.0
flake8==7.1.1

